move models folder
