+++ src/controllers/roles.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import Role from '../database/models/role';
+import Role from '../models/role';
 
 async function index(req: Request, res: Response) {
   res.send('Role controller');
