+++ src/controllers/users.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from 'express';
 import { validationResult } from 'express-validator';
-import db from '../database/models/index';
+import db from '../models/index';
 
 const getAll = async (req: Request, res: Response) => {
   const errors = validationResult(req);
