+++ src/models/slide.ts
