+++ src/services/categoryService.ts
@@ -1,4 +1,4 @@
-import db from '../database/models';
+import db from '../models';
 
 export async function getCategory(id: number) {
   return db.Category.findByPk(id);
